updated to address runtime error

+++ M7. NLP/Week 1 - Traditional NLP/3. Chatbot/my_chatbot/train.py
@@ -78,7 +78,7 @@ optimizer = torch.optim.Adam(model.parameters(), lr = learning_rate)
 for epoch in range(num_epochs):
     for (words, labels) in train_loader:
         words = words.to(device)
-        labels = labels.to(device)
+        labels = labels.to(device, dtype=torch.int64)
         
         # Forward pass
         outputs = model(words)
